Refactor: 자동식단, 수동식단, 수동식단수정 페이지 -  useFetchMinorCategories 훅 분리
