+++ src/components/feature/MenualPlan/index.tsx
@@ -2,7 +2,6 @@
 
 import { useEffect, useState } from 'react';
 import { zodResolver } from '@hookform/resolvers/zod';
-import { useQueryClient } from '@tanstack/react-query';
 import { useForm } from 'react-hook-form';
 import { mealHeaderSchema } from '@/schema/mealSchema';
 import { CalendarInfo } from '@/type/mealType';
@@ -11,26 +10,24 @@ import {
   HandleChangeCategoryParam,
   SelectedCategory,
 } from '@/type/menuCategory/category';
-import { Result } from '@/type/response';
 import {
   getCurrentYearMonthNow,
-  hasNonEmptyFoods,
+  hasFoods,
   isValidDateString,
 } from '@/utils/calendar';
 import InfoCard from '@/components/common/InfoCard';
 import MealForm from '@/components/common/MealForm';
-import { Option } from '@/components/common/Selectbox';
 import MealCalendar from '@/components/shared/Meal/MealCalender';
 import MealHeader, {
   MealHeaderFormData,
 } from '@/components/shared/Meal/MealHeader';
 import { INFOCARD_MESSAGE } from '@/constants/_infoCard';
-import { MAJOR_CATEGORIES } from '@/constants/_meal';
 import { MEAL_FORM_LEGEND } from '@/constants/_MealForm';
 import { ROUTES } from '@/constants/_navbar';
 import { PAGE_TITLE } from '@/constants/_pageTitle';
 import { MEAL_HEADER_ERROR } from '@/constants/_schema';
 import { MEAL_CREATE_MESSAGE } from '@/constants/_toastMessage';
+import { useFetchMinorCategories } from '@/hooks/menuCategory/useFetchMinorCategories';
 import { usePrefetchMinorCategories } from '@/hooks/menuCategory/usePrefetchMinorCategories';
 import useNavigate from '@/hooks/useNavigate';
 import { useMenualPlanStore } from '@/stores/useMenualPlanStore';
@@ -45,7 +42,6 @@ const MenualPlan = () => {
   });
   const { year, month } = getCurrentYearMonthNow();
   const [isCategoryError, setIsCategoryError] = useState(false);
-  const [minorCategories, setMinorCategories] = useState<Option[]>([]);
   const showToast = useToastStore((state) => state.showToast);
   const { setMonthMenuName, setCategory, setCalendar } = useMenualPlanStore(
     (state) => ({
@@ -56,7 +52,9 @@ const MenualPlan = () => {
   );
   const { navigate } = useNavigate();
 
-  const queryClient = useQueryClient();
+  const { minorCategories } = useFetchMinorCategories(
+    selectedCategory.majorCategory,
+  );
   const { prefetchMinorCategories, hasCategories } =
     usePrefetchMinorCategories();
 
@@ -111,7 +109,7 @@ const MenualPlan = () => {
       return;
     }
 
-    if (!hasNonEmptyFoods(calendarData)) {
+    if (!hasFoods(calendarData)) {
       showToast(MEAL_CREATE_MESSAGE.error.emptyFood, 'warning', 3000);
     } else {
       setMonthMenuName(data.monthMenuName);
@@ -129,41 +127,40 @@ const MenualPlan = () => {
   };
 
   useEffect(() => {
-    if (!hasCategories) {
-      prefetchMinorCategories();
-    }
+    if (hasCategories) return;
+    prefetchMinorCategories();
   }, [hasCategories, prefetchMinorCategories]);
 
-  useEffect(() => {
-    const fetchCategories = () => {
-      switch (selectedCategory.majorCategory) {
-        case MAJOR_CATEGORIES[0]:
-          return queryClient.getQueryData<Result<string[]>>([
-            'getSchoolMinorCategories',
-          ]);
-        case MAJOR_CATEGORIES[1]:
-          return queryClient.getQueryData<Result<string[]>>([
-            'getSchoolNameMinorCategories',
-          ]);
-        case MAJOR_CATEGORIES[2]:
-          return queryClient.getQueryData<Result<string[]>>([
-            'getHospitalMinorCategories',
-          ]);
-        default:
-          return null;
-      }
-    };
+  // useEffect(() => {
+  //   const fetchCategories = () => {
+  //     switch (selectedCategory.majorCategory) {
+  //       case MAJOR_CATEGORIES[0]:
+  //         return queryClient.getQueryData<Result<string[]>>([
+  //           'getSchoolMinorCategories',
+  //         ]);
+  //       case MAJOR_CATEGORIES[1]:
+  //         return queryClient.getQueryData<Result<string[]>>([
+  //           'getSchoolNameMinorCategories',
+  //         ]);
+  //       case MAJOR_CATEGORIES[2]:
+  //         return queryClient.getQueryData<Result<string[]>>([
+  //           'getHospitalMinorCategories',
+  //         ]);
+  //       default:
+  //         return null;
+  //     }
+  //   };
 
-    const categories = fetchCategories();
-    if (!categories) return;
-    const { data } = categories;
-    if (!data) return;
-    const formattedData = data.map((category) => ({
-      value: category,
-      label: category,
-    }));
-    setMinorCategories(formattedData);
-  }, [selectedCategory.majorCategory]);
+  //   const categories = fetchCategories();
+  //   if (!categories) return;
+  //   const { data } = categories;
+  //   if (!data) return;
+  //   const formattedData = data.map((category) => ({
+  //     value: category,
+  //     label: category,
+  //   }));
+  //   setMinorCategories(formattedData);
+  // }, [selectedCategory.majorCategory]);
 
   return (
     <div className='flex gap-8'>
