+++ src/components/feature/MenualPlanEdit/index.tsx
@@ -1,8 +1,7 @@
 'use client';
 
-import { useEffect, useState } from 'react';
+import { useState } from 'react';
 import { zodResolver } from '@hookform/resolvers/zod';
-import { useQueryClient } from '@tanstack/react-query';
 import { AxiosError } from 'axios';
 import { useForm } from 'react-hook-form';
 import { mealHeaderSchema } from '@/schema/mealSchema';
@@ -20,15 +19,14 @@ import {
   transformCalendarToPostSave,
 } from '@/utils/calendar';
 import MealForm from '@/components/common/MealForm';
-import { Option } from '@/components/common/Selectbox';
 import MealCalendar from '@/components/shared/Meal/MealCalender';
 import MealHeader from '@/components/shared/Meal/MealHeader';
-import { MAJOR_CATEGORIES } from '@/constants/_meal';
 import { MEAL_FORM_LEGEND } from '@/constants/_MealForm';
 import { ROUTES } from '@/constants/_navbar';
 import { PAGE_TITLE } from '@/constants/_pageTitle';
 import { MEAL_HEADER_ERROR } from '@/constants/_schema';
 import { usePostMonthMenusSave } from '@/hooks/menu/usePostMonthMenusSave';
+import { useFetchMinorCategories } from '@/hooks/menuCategory/useFetchMinorCategories';
 import useNavigate from '@/hooks/useNavigate';
 import { useMenualPlanStore } from '@/stores/useMenualPlanStore';
 import { useToastStore } from '@/stores/useToastStore';
@@ -45,11 +43,12 @@ const MenualPlanEdit = () => {
     useState<SelectedCategory>(category);
   const { year, month } = getCurrentYearMonthNow();
   const [isCategoryError, setIsCategoryError] = useState(false);
-  const [minorCategories, setMinorCategories] = useState<Option[]>([]);
   const showToast = useToastStore((state) => state.showToast);
   const { navigate } = useNavigate();
 
-  const queryClient = useQueryClient();
+  const { minorCategories } = useFetchMinorCategories(
+    selectedCategory.majorCategory,
+  );
   const { mutate: postSaveMutate } = usePostMonthMenusSave();
 
   const {
@@ -144,37 +143,6 @@ const MenualPlanEdit = () => {
       return;
     }
   };
-
-  useEffect(() => {
-    const fetchCategories = () => {
-      switch (selectedCategory.majorCategory) {
-        case MAJOR_CATEGORIES[0]:
-          return queryClient.getQueryData<Result<string[]>>([
-            'getSchoolMinorCategories',
-          ]);
-        case MAJOR_CATEGORIES[1]:
-          return queryClient.getQueryData<Result<string[]>>([
-            'getSchoolNameMinorCategories',
-          ]);
-        case MAJOR_CATEGORIES[2]:
-          return queryClient.getQueryData<Result<string[]>>([
-            'getHospitalMinorCategories',
-          ]);
-        default:
-          return null;
-      }
-    };
-
-    const categories = fetchCategories();
-    if (!categories) return;
-    const { data } = categories;
-    if (!data) return;
-    const formattedData = data.map((category) => ({
-      value: category,
-      label: category,
-    }));
-    setMinorCategories(formattedData);
-  }, [selectedCategory.majorCategory]);
 
   return (
     <MealForm
